fix by code review
