+++ src/components/Details/Details.js
@@ -56,7 +56,6 @@ import {
 import { isEveryObjectValueEmpty } from '../../utils/isEveryObjectValueEmpty'
 import { showArtifactsPreview } from '../../reducers/artifactsReducer'
 import { setFieldState } from 'igz-controls/utils/form.util'
-import { datePickerPastOptions, PAST_24_HOUR_DATE_OPTION } from '../../utils/datePicker.util'
 
 import './details.scss'
 
@@ -79,7 +78,6 @@ const Details = ({
   setChanges,
   setChangesCounter,
   setChangesData,
-  setDetailsDates,
   setInfoContent,
   setIteration,
   setIterationOption,
@@ -249,30 +247,6 @@ const Details = ({
     setFiltersWasHandled
   ])
 
-  const handleChangeDates = useCallback(
-    (dates, isPredefined, selectedOptionId) => {
-      const generatedDates = [...dates]
-
-      if (generatedDates.length === 1) {
-        generatedDates.push(new Date())
-      }
-      setDetailsDates({
-        value: generatedDates,
-        selectedOptionId,
-        isPredefined
-      })
-    },
-    [setDetailsDates]
-  )
-
-  useEffect(() => {
-    const past24hoursOption = datePickerPastOptions.find(
-      option => option.id === PAST_24_HOUR_DATE_OPTION
-    )
-
-    handleChangeDates(past24hoursOption.handler(), true, PAST_24_HOUR_DATE_OPTION)
-  }, [handleChangeDates])
-
   return (
     <Form form={formRef.current} onSubmit={() => {}}>
       {formState => (
@@ -288,7 +262,6 @@ const Details = ({
               getCloseDetailsLink={getCloseDetailsLink}
               isDetailsScreen={isDetailsScreen}
               handleCancel={handleCancel}
-              handleChangeDates={handleChangeDates}
               handleRefresh={handleRefresh}
               handleShowWarning={handleShowWarning}
               pageData={pageData}
@@ -302,7 +275,6 @@ const Details = ({
             <DetailsTabsContent
               applyChangesRef={applyChangesRef}
               formState={formState}
-              handleChangeDates={handleChangeDates}
               handlePreview={handlePreview}
               pageData={pageData}
               selectedItem={selectedItem}
@@ -311,7 +283,6 @@ const Details = ({
               setChangesData={setChangesData}
               setIteration={setIteration}
               setIterationOption={setIterationOption}
-              setSelectedMetricsOptions={setSelectedMetricsOptions}
             />
           </div>
           {blocker.state === 'blocked' && (
